refactor: move fetch-latest scheduler logic to task module
